+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add custom data uploads via `ga_custom_upload_file()`
 * Cleaned up documentation a bit
+* Added auto-authentication if you specify environment var `GA_AUTH_FILE`
 
 # 0.3.0 CRAN